Add "tproxy" option Added the function of "MacOS" FreeBSD firewall traffic forwarding and resolving destination address example:   "inbounds": [     {       "listen": "127.0.0.1",       "port": 1122,       "protocol": "dokodemo-door",       "tag": "dokodemo",       "settings": {         "network": "tcp",         "followRedirect": true,         "userLevel": 0       },       "streamSettings": {         "sockopt": {           "tproxy": "pf"         }       }     }   ]
